feat: add TwoLevelPieChart example in TypeScript
